+++ packages/enhanced/src/lib/container/runtime/FederationRuntimePlugin.ts
@@ -12,7 +12,7 @@ import {
 import fs from 'fs';
 import path from 'path';
 import { TEMP_DIR } from '../constant';
-import type { ModuleFederationPluginOptions } from '../../../declarations/plugins/container/ModuleFederationPlugin';
+import type { moduleFederationPlugin } from '@module-federation/sdk';
 import HoistContainerReferences from '../HoistContainerReferencesPlugin';
 
 const { RuntimeGlobals, Template } = require(
@@ -37,13 +37,13 @@ const RuntimePath = require.resolve('@module-federation/runtime', {
 const federationGlobal = getFederationGlobalScope(RuntimeGlobals);
 
 class FederationRuntimePlugin {
-  options?: ModuleFederationPluginOptions;
+  options?: moduleFederationPlugin.ModuleFederationPluginOptions;
   entryFilePath: string;
   pluginsFilePath: string; // New path for plugins file
   bundlerRuntimePath: string;
 
-  constructor(options?: ModuleFederationPluginOptions) {
-    this.options = options;
+  constructor(options?: moduleFederationPlugin.ModuleFederationPluginOptions) {
+    this.options = options ? { ...options } : undefined;
     this.entryFilePath = '';
     this.pluginsFilePath = ''; // Initialize plugins file path
     this.bundlerRuntimePath = BundlerRuntimePath;