+++ packages/enhanced/test/configCases/sharing/consume-multiple-versions/index.js
@@ -174,7 +174,6 @@ it('should be able to consume different shared module version depending on conte
   );
   expect(require('my-module4')).toBe('shared@9.9.9');
   // expectWarning();
-  debugger;
   expect(require('shared2')).toBe('shared2@9.9.9');
   // expectWarning(
   //   /No satisfying version \(=1\.2\.3 =3\.2\.1\) of shared module shared2 found in shared scope default/,