+++ packages/sdk/src/types/common.ts
@@ -10,9 +10,3 @@ export interface RemoteWithVersion {
 
 export type RemoteEntryInfo = RemoteWithEntry | RemoteWithVersion;
 export type Module = any;
-
-declare namespace NodeJS {
-  interface ProcessEnv {
-    FEDERATION_WEBPACK_PATH?: string;
-  }
-}