+++ packages/sdk/src/types/stats.ts
@@ -1,3 +1,4 @@
+import { MFModuleType } from '../constant';
 import type { RemoteWithEntry, RemoteWithVersion } from './common';
 
 export type RemoteEntryType = 'esm' | 'global';
@@ -18,8 +19,11 @@ export interface BasicStatsMetaData {
   globalName: string;
   buildInfo: StatsBuildInfo;
   remoteEntry: ResourceInfo;
+  prefetchEntry?: ResourceInfo;
   prefetchInterface: boolean;
   types: Omit<ResourceInfo, 'type'>;
+  type: string;
+  pluginVersion: string;
 }
 
 type StatsMetaDataWithGetPublicPath<T = BasicStatsMetaData> = T & {
@@ -75,7 +79,7 @@ export type StatsRemote<T = StatsRemoteVal> =
 
 export interface StatsModuleInfo {
   name: string;
-  file: string;
+  file: string[];
 }
 
 export interface ManifestModuleInfos {
@@ -85,7 +89,7 @@ export interface ManifestModuleInfos {
 export interface StatsExpose {
   id: string;
   name: string;
-  path?: string; // 对线上已有数据兼容过渡至 2023-05-26
+  path?: string;
   file: string;
   requires: string[];
   assets: StatsAssets;